Create a <Fingering /> component
