+++ src/lib/components/GuitarChord/Fingering.js
@@ -1,0 +1,53 @@
+import React from 'react';
+import CONSTANTS from './constants';
+
+const Fingering = ({ frets, fingering, numberFretsToRender, lowestFret }) => {
+  // this component draws the black circles representing fingered notes, if
+  // this.props.fingering is passed into the GuitarChord, it will also put the
+  // numbers in the circles that represent which finger to use
+  const svgFingering = [];
+  // TODO: this is duplicate code in other render functions
+  const width = CONSTANTS.X_MAX - CONSTANTS.X_BASE;
+  const offset = width / 5;
+  const isOpenChord = frets.includes(0);
+  const yOffset = CONSTANTS.Y_MAX / numberFretsToRender;
+
+  if (!isOpenChord) {
+    // add a fret label to the left of the chord diagram
+    svgFingering.push(
+      <text key={'todo'} x={CONSTANTS.X_MAX + 10} y={CONSTANTS.Y_BASE} fill="black">
+      {lowestFret}
+      </text>
+    );
+  }
+
+  frets.forEach((fret, index) => {
+    if (fret === 'x' || fret === 0) {
+      const xValue = CONSTANTS.X_BASE + (offset * index) - 4; // -1 because of how text sits
+      return svgFingering.push(
+        <text x={xValue} y={CONSTANTS.Y_BASE - 3 }>{fret}</text>
+      );
+    }
+    const xValue = CONSTANTS.X_BASE + (offset * index);
+    let yCoord;
+    if (isOpenChord) {
+      yCoord = CONSTANTS.Y_BASE + (yOffset * fret) - (yOffset / 2);
+    } else {
+      yCoord = CONSTANTS.Y_BASE + (yOffset * (fret - lowestFret + 1)) - (yOffset / 2);
+    }
+    svgFingering.push(
+      <circle cx={xValue} cy={yCoord} r="8" fill="black"></circle>
+    );
+    if (fingering) {
+      const finger = fingering[index];
+      svgFingering.push(
+        <text x={xValue} y={yCoord + 5} fill="white" textAnchor="middle">
+        {finger}
+        </text>
+      );
+    }
+  });
+  return svgFingering;
+};
+
+export default Fingering;
